Fix AppShelf circular imports
